requirements.txt
django==1.7
mysql-python
django-suit==0.2.18
django-debug-toolbar==1.3.2
pytils==0.3
pytz
pillow
xlwt==1.0.0
pyexcel==0.2.2
pyexcel-xls
pyexcel-xlsx
django-annoying==0.8.3
django-imagekit==3.2.7
django-widget-tweaks==1.4.1
django-ckeditor==4.5.1
django-cleanup==0.4.2
djangorestframework==3.3.3
python-dateutil==2.6.0